refactor(core): move some documentPerspective components from structure to core
