+++ packages/sanity/src/core/index.ts
@@ -24,29 +24,20 @@ export * from './presence'
 export * from './preview'
 export {
   AddedVersion,
-  CreatedRelease,
-  createReleaseId,
-  DEFAULT_RELEASE_TYPE,
   DiscardVersionDialog,
-  type EditableReleaseDocument,
   getBundleIdFromReleaseDocumentId,
-  getCreateVersionOrigin,
-  getDocumentIsInPerspective,
   getPublishDateFromRelease,
   getReleaseTone,
   isDraftPerspective,
   isPublishedPerspective,
   isReleaseDocument,
   LATEST,
-  ReleaseAvatar,
-  ReleaseBadge,
   type ReleaseDocument,
-  ReleaseForm,
   useDocumentVersions,
   usePerspective,
-  useReleaseOperations,
   useReleases,
   useVersionOperations,
+  VersionChip,
   versionDocumentExists,
 } from './releases'
 export * from './scheduledPublishing'
