+++ packages/sanity/src/core/releases/components/index.ts
@@ -1,4 +1,4 @@
 export * from './dialog/DiscardVersionDialog'
 export * from './dialog/ReleaseForm'
+export * from './documentHeader/VersionChip'
 export * from './ReleaseAvatar'
-export * from './ReleaseBadge'
