+++ packages/sanity/src/structure/panes/document/documentPanel/header/perspective/__tests__/DocumentHeaderTitle.test.tsx
@@ -36,10 +36,8 @@ vi.mock('sanity', async (importOriginal) => {
     ...(await importOriginal()),
     unstable_useValuePreview: vi.fn(),
     useDocumentVersions: vi.fn(),
-    useDateTimeFormat: vi.fn().mockReturnValue({format: vi.fn()}),
     useReleases: vi.fn(),
     usePerspective: vi.fn(() => ({perspective: undefined})),
-    useVersionOperations: vi.fn(() => ({})),
   }
 })
